@@ -1,37 +1,35 @@
 class PostsController < ApplicationController
-  
-  before_filter :authenticate, :except => [:index, :show]
 
   # GET /posts
-  # GET /posts.xml
+  # GET /posts.json
   def index
-    @posts = Post.all
+    @posts = Post.paginate(:page => params[:page], :per_page => 5)
 
     respond_to do |format|
       format.html # index.html.erb
-      format.xml  { render :xml => @posts }
+      format.json { render json: @posts }
     end
   end
 
   # GET /posts/1
-  # GET /posts/1.xml
+  # GET /posts/1.json
   def show
     @post = Post.find(params[:id])
 
     respond_to do |format|
       format.html # show.html.erb
-      format.xml  { render :xml => @post }
+      format.json { render json: @post }
     end
   end
 
   # GET /posts/new
-  # GET /posts/new.xml
+  # GET /posts/new.json
   def new
     @post = Post.new
 
     respond_to do |format|
       format.html # new.html.erb
-      format.xml  { render :xml => @post }
+      format.json { render json: @post }
     end
   end
 
@@ -41,46 +39,46 @@ class PostsController < ApplicationController
   end
 
   # POST /posts
-  # POST /posts.xml
+  # POST /posts.json
   def create
     @post = Post.new(params[:post])
 
     respond_to do |format|
       if @post.save
-        format.html { redirect_to(@post, :notice => 'Post was successfully created.') }
-        format.xml  { render :xml => @post, :status => :created, :location => @post }
+        format.html { redirect_to @post, notice: 'Post was successfully created.' }
+        format.json { render json: @post, status: :created, location: @post }
       else
-        format.html { render :action => "new" }
-        format.xml  { render :xml => @post.errors, :status => :unprocessable_entity }
+        format.html { render action: "new" }
+        format.json { render json: @post.errors, status: :unprocessable_entity }
       end
     end
   end
 
   # PUT /posts/1
-  # PUT /posts/1.xml
+  # PUT /posts/1.json
   def update
     @post = Post.find(params[:id])
 
     respond_to do |format|
       if @post.update_attributes(params[:post])
-        format.html { redirect_to(@post, :notice => 'Post was successfully updated.') }
-        format.xml  { head :ok }
+        format.html { redirect_to @post, notice: 'Post was successfully updated.' }
+        format.json { head :ok }
       else
-        format.html { render :action => "edit" }
-        format.xml  { render :xml => @post.errors, :status => :unprocessable_entity }
+        format.html { render action: "edit" }
+        format.json { render json: @post.errors, status: :unprocessable_entity }
       end
     end
   end
 
   # DELETE /posts/1
-  # DELETE /posts/1.xml
+  # DELETE /posts/1.json
   def destroy
     @post = Post.find(params[:id])
     @post.destroy
 
     respond_to do |format|
-      format.html { redirect_to(posts_url) }
-      format.xml  { head :ok }
+      format.html { redirect_to posts_url }
+      format.json { head :ok }
     end
   end
 end
