@@ -1,17 +1,17 @@
 class CommentsController < ApplicationController
-  
-  before_filter :authenticate, :only => :destroy
 
-  def create
-    @post = Post.find(params[:post_id])
-    @comment = @post.comments.create(params[:comment])
-    redirect_to post_path(@post)
-  end
-  
-  def destroy
-    @post = Post.find(params[:post_id])
-    @comment = @post.comments.find(params[:id])
-    @comment.destroy
-    redirect_to post_path(@post)
-  end  	
+	http_basic_authenticate_with :name => 'admin', :password => 'root', :only => :destroy
+
+	def create
+		@post = Post.find(params[:post_id])
+		@comment = @post.comments.create(params[:comment])
+		redirect_to post_path(@post)
+	end
+
+	def destroy
+		@post = Post.find(params[:post_id])
+		@comment = @post.comments.find(params[:id])
+		@comment.destroy
+		redirect_to post_path(@post)
+	end
 end
