@@ -1,5 +1,5 @@
 module PostsHelper
-  def join_tags(post)
-    post.tags.map { |t| t.name }.join(", ")
-  end
+		def join_tags(post)
+			post.tags.map { |t| t.name }.join(", ")
+		end
 end
