@@ -4,4 +4,4 @@
 # If you change this key, all old signed cookies will become invalid!
 # Make sure the secret is at least 30 characters and all random,
 # no regular words or you'll be exposed to dictionary attacks.
-Blog::Application.config.secret_token = '7b0cdb7934970fce844fe18eda371b8ec1a9d3e3f7c1397ccaef5d14f968ae563212d14068826bc1193714ad74b9584f1d51df8493d3713b8c9bdfcf97dbf0b6'
+Blog::Application.config.secret_token = '5af2d6689bef3ab406f931e5181b97340df56cb35ec249050c57eac8d6df0454433447874e5686dd7c1bb47734f70911e40a26fc7ae12597bd8bb47cd48b0d3c'
