@@ -1,6 +1,6 @@
 Blog::Application.routes.draw do
   resources :posts do
-    resources :comments
+      resources :comments
   end
 
   get "home/index"
@@ -54,7 +54,7 @@ Blog::Application.routes.draw do
 
   # You can have the root of your site routed with "root"
   # just remember to delete public/index.html.
-	root :to => "home#index"
+  root :to => 'home#index'
 
   # See how all your routes lay out with "rake routes"
 
