@@ -1,5 +1,5 @@
 class CreatePosts < ActiveRecord::Migration
-  def self.up
+  def change
     create_table :posts do |t|
       t.string :name
       t.string :title
@@ -7,9 +7,5 @@ class CreatePosts < ActiveRecord::Migration
 
       t.timestamps
     end
-  end
-
-  def self.down
-    drop_table :posts
   end
 end
