@@ -1,5 +1,5 @@
 class CreateComments < ActiveRecord::Migration
-  def self.up
+  def change
     create_table :comments do |t|
       t.string :commenter
       t.text :body
@@ -7,9 +7,6 @@ class CreateComments < ActiveRecord::Migration
 
       t.timestamps
     end
-  end
-
-  def self.down
-    drop_table :comments
+    add_index :comments, :post_id
   end
 end
