@@ -1,14 +1,11 @@
 class CreateTags < ActiveRecord::Migration
-  def self.up
+  def change
     create_table :tags do |t|
       t.string :name
       t.references :post
 
       t.timestamps
     end
-  end
-
-  def self.down
-    drop_table :tags
+    add_index :tags, :post_id
   end
 end
