@@ -1,3 +1,4 @@
+# encoding: UTF-8
 # This file is auto-generated from the current state of the database. Instead
 # of editing this file, please use the migrations feature of Active Record to
 # incrementally modify your database, and then regenerate this schema definition.
@@ -10,7 +11,7 @@
 #
 # It's strongly recommended to check this file into your version control system.
 
-ActiveRecord::Schema.define(:version => 20110416232852) do
+ActiveRecord::Schema.define(:version => 20111106220800) do
 
   create_table "comments", :force => true do |t|
     t.string   "commenter"
@@ -19,6 +20,8 @@ ActiveRecord::Schema.define(:version => 20110416232852) do
     t.datetime "created_at"
     t.datetime "updated_at"
   end
+
+  add_index "comments", ["post_id"], :name => "index_comments_on_post_id"
 
   create_table "posts", :force => true do |t|
     t.string   "name"
@@ -34,5 +37,7 @@ ActiveRecord::Schema.define(:version => 20110416232852) do
     t.datetime "created_at"
     t.datetime "updated_at"
   end
+
+  add_index "tags", ["post_id"], :name => "index_tags_on_post_id"
 
 end
